test(portfolio-contract): split portfolio.contract.test into 5 files
AVA parallelizes across test *files* (one worker per file), not within a
file. All 48 tests of portfolio.contract.test.ts ran in a single worker
— one CPU core — making it the ~24min long pole of test-portfolio-contract
while the runner's other 3 cores sat idle after the other suites drained.

Split the suite by feature area so AVA can spread it across cores, with
shared helpers extracted verbatim to contract-test-support.ts:

  core         13 tests  ~17s   open/deposit/withdraw, USDN/Aave/Compound
  vaults        9 tests  ~19s   Beefy/ERC4626 positions
  evm          12 tests  ~10s   EVM account opening, config, recovery
  evm-handler   6 tests  ~21s   evmHandler.* operations
  planner       8 tests  ~11s   planner-driven flows (incl. the O(N^2) ones)

Balanced so no file dominates (single-worker times above; the original
ran ~93s of work serialized on one core). On the 4-vCPU ubuntu-latest
runner the critical path drops from the whole-file time to ~max(file),
roughly a 4-5x reduction for this package. evm was split from evm-handler
after measuring the combined file at ~47s — 2.5x the others.

No behavior change: 48 tests + 1 todo preserved; full package suite still
339 passed / 11 skipped / 8 todo. Snapshots regenerated per file.
